requirements.txt
# ================================================================================================
# TRACKPRO COMPLETE DEPENDENCY LIST - GITHUB DEPLOYMENT READY
# For GitHub deployment - ensures TrackPro works on ANY computer
# ================================================================================================

# CORE FRAMEWORK DEPENDENCIES
PyQt6>=6.4.0,<6.7.0            # Main GUI framework
PyQt6-WebEngine>=6.4.0,<6.7.0  # Web engine for OAuth/Discord integration  
PyQt6-Charts>=6.4.0,<6.7.0     # Charts and graphs for telemetry visualization
numpy>=1.19.0,<2.0.0           # Scientific computing foundation
requests>=2.25.0,<3.0.0        # HTTP requests for APIs
psutil>=5.9.0,<6.0.0           # System resource monitoring

# ================================================================================================
# IRACING & SIMULATION DEPENDENCIES  
# ================================================================================================
pyirsdk>=1.2.0                 # iRacing SDK interface
pygame>=2.0.0,<3.0.0           # Audio and input handling
PyYAML>=5.4.0,<7.0.0           # YAML parsing for iRacing session data

# ================================================================================================
# WINDOWS SYSTEM DEPENDENCIES (Required on Windows)
# ================================================================================================
pywin32>=300                   # Windows API access (services, registry, etc.)
pywin32-ctypes>=0.2.0          # Windows system calls

# ================================================================================================
# DATABASE & CLOUD DEPENDENCIES
# ================================================================================================
supabase>=0.7.1,<2.0.0         # Cloud database and authentication
python-dotenv>=0.19.0,<2.0.0   # Environment variable management

# ================================================================================================
# SMS 2FA DEPENDENCIES
# ================================================================================================
twilio>=8.5.0,<9.0.0           # SMS and phone verification for 2FA

# ================================================================================================
# DATA VISUALIZATION & ANALYSIS
# ================================================================================================
matplotlib>=3.5.0,<4.0.0       # Plotting and track visualization
pyqtgraph>=0.13.0,<1.0.0       # Real-time plotting for telemetry
scipy>=1.7.0,<2.0.0            # Scientific computing (track interpolation)

# ================================================================================================
# ADVANCED TRACK BUILDING DEPENDENCIES (Machine Learning)
# ================================================================================================
scikit-learn>=1.0.0,<2.0.0     # Machine learning for track analysis
filterpy>=1.4.0,<2.0.0         # Kalman filtering for data smoothing

# ================================================================================================
# EYE TRACKING DEPENDENCIES (Advanced feature)
# ================================================================================================
eyetrax>=0.3.1,<1.0.0          # Eye tracking library
mediapipe==0.10.21             # Computer vision for eye tracking (specific version required)

# ================================================================================================
# AI COACH DEPENDENCIES (OpenAI & ElevenLabs)
# ================================================================================================
openai>=1.86.0,<2.0.0          # OpenAI API for coaching advice
elevenlabs>=0.2.27,<1.0.0      # Text-to-speech for AI coach
soundfile>=0.12.1,<1.0.0       # Audio file handling
pydub>=0.25.1,<1.0.0           # Audio processing

# ================================================================================================
# PAYMENT & GAMIFICATION DEPENDENCIES
# ================================================================================================
stripe>=5.0.0,<6.0.0           # Payment processing for Race Pass

# ================================================================================================
# PACKAGING & DISTRIBUTION 
# ================================================================================================
PyInstaller>=6.0.0,<7.0.0      # For creating standalone executables

# ================================================================================================
# ENHANCED FEATURES & NETWORKING
# ================================================================================================
python-socketio>=5.8.0,<6.0.0  # Real-time communication
redis>=4.5.0,<5.0.0            # Caching and session management
pillow>=9.5.0,<11.0.0          # Image processing
opencv-python>=4.7.0,<5.0.0    # Computer vision
cryptography>=40.0.0,<42.0.0   # Encryption and security
pydantic>=1.10.0,<3.0.0        # Data validation
httpx>=0.24.0,<1.0.0           # Modern HTTP client
websockets>=11.0.0,<12.0.0     # WebSocket communication
plyer>=2.1.0,<3.0.0            # Cross-platform notifications

# ================================================================================================
# BACKGROUND PROCESSING & MONITORING
# ================================================================================================
celery>=5.2.0,<6.0.0           # Background task processing
sentry-sdk>=1.25.0,<2.0.0      # Error monitoring and reporting

# ================================================================================================
# UTILITY & COMPATIBILITY
# ================================================================================================
packaging>=21.0,<24.0          # Package version handling
typing-extensions>=4.0.0,<5.0.0 # Type hints for older Python versions

# ================================================================================================
# OPTIONAL AUDIO ENHANCEMENTS
# ================================================================================================
sounddevice>=0.4.0,<1.0.0      # Advanced audio I/O (notifications, AI coach)

# ================================================================================================
# INSTALLATION NOTES FOR USERS
# ================================================================================================
# 
# AUTOMATIC INSTALLATION:
# ----------------------
# Run: pip install --user -r requirements.txt
# 
# SYSTEM REQUIREMENTS:
# -------------------
# - Windows 10/11 (64-bit)
# - Python 3.8+ (3.11 recommended for eye tracking)
# - 4GB RAM minimum, 8GB recommended
# - 500MB free disk space
# - Internet connection for cloud features
# 
# OPTIONAL COMPONENTS:
# -------------------
# - vJoy: For pedal functionality (included in installer)
# - HidHide: For device hiding (included in installer)  
# - Visual C++ Redistributable (included in installer)
# 
# TROUBLESHOOTING:
# ---------------
# If installation fails:
# 1. Update pip: python -m pip install --upgrade pip
# 2. Use --user flag: pip install --user -r requirements.txt
# 3. Try: pip install --user --only-binary=all -r requirements.txt
# 4. For eye tracking: Use Python 3.11 specifically
# 5. If specific packages fail: pip install --user <package-name>
# 6. Common Windows issues:
#    - Install Visual C++ Build Tools if needed
#    - Use Windows Store Python if available
#    - Try: pip install --user --upgrade setuptools wheel
# 7. If pywin32 fails: pip install --user pywin32 --force-reinstall
# 
# ================================================================================================
